app/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
mongoengine>=0.24.0
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
httpx>=0.23.0
geopy>=2.2.0

timezonefinder>=6.2.0
azure-appconfiguration-provider>=1.2.0

gunicorn
